requirements.txt
aiogram==3.21.0
requests==2.32.4
python-dotenv==1.1.1
aiohttp==3.12.14
pydantic==2.11.7
